Script for an IRC bot that deals with last.fm
The bot is currently capable of showing a user's now playing track and
comparing the tastes between two users.

Don't mind the hardcoded %nick_user_map, it will be made actually
customizable, with IRC-accessible commands.

It also has a feature to show the currently playing track of all users
in the channel that are listening on last.fm, but presently the bot will
only obey the command if the bot "itself" says it.